# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  Asset:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      environments:
        additionalProperties:
          $ref: '#/definitions/Environment'
        type: object
      external_links:
        items:
          $ref: '#/definitions/AssetExternalLink'
        type: array
      has_run_history:
        type: boolean
      id:
        type: string
      is_stub:
        type: boolean
      last_sync_at:
        type: string
      metadata:
        additionalProperties: true
        type: object
      mrn:
        type: string
      name:
        type: string
      parent_mrn:
        type: string
      providers:
        items:
          type: string
        type: array
      query:
        type: string
      query_language:
        type: string
      schema:
        additionalProperties:
          type: string
        type: object
      sources:
        items:
          $ref: '#/definitions/AssetSource'
        type: array
      tags:
        items:
          type: string
        type: array
      terms:
        description: |-
          Terms carries glossary term names a discovery assigned to this
          asset. It is input only: the links live in asset_terms, so reading
          an asset back from the database never fills this in.
        items:
          type: string
        type: array
      type:
        type: string
      updated_at:
        type: string
      user_description:
        type: string
    type: object
...
  AssetExternalLink:
    properties:
      icon:
        type: string
      name:
        type: string
      url:
        type: string
    type: object
...
  AssetSource:
    properties:
      last_sync_at:
        type: string
      name:
        type: string
      priority:
        type: integer
      properties:
        additionalProperties: true
        type: object
    type: object
...
  BatchLineageResult:
    properties:
      edge:
        $ref: '#/definitions/LineageEdge'
      status:
        description: '"created", "duplicate", or "existing"'
        type: string
    type: object
...
  Dataset:
    properties:
      facets:
        additionalProperties: true
        type: object
      inputFacets:
        additionalProperties: true
        type: object
      name:
        type: string
      namespace:
        type: string
      outputFacets:
        additionalProperties: true
        type: object
    type: object
...
  Environment:
    properties:
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      path:
        type: string
    type: object
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  Job:
    properties:
      facets:
        additionalProperties: true
        type: object
      name:
        type: string
      namespace:
        type: string
    type: object
...
  LineageEdge:
    properties:
      id:
        type: string
      job_mrn:
        type: string
      last_seen_at:
        type: string
      observation_count:
        type: integer
      origin:
        type: string
      source:
        type: string
      target:
        type: string
      type:
        type: string
    type: object
  LineageNode:
    properties:
      asset:
        $ref: '#/definitions/Asset'
      depth:
        type: integer
      id:
        type: string
      type:
        type: string
    type: object
  LineageResponse:
    properties:
      edges:
        items:
          $ref: '#/definitions/LineageEdge'
        type: array
      nodes:
        items:
          $ref: '#/definitions/LineageNode'
        type: array
    type: object
...
  LineageRun:
    properties:
      facets:
        additionalProperties: true
        type: object
      runId:
        type: string
    type: object
...
  ReindexAcceptedResponse:
    properties:
      message:
        example: Reindex started
        type: string
      status:
        example: accepted
        type: string
    type: object
  ReindexStatusResponse:
    properties:
      es_configured:
        type: boolean
      running:
        type: boolean
    type: object
...
  RunEvent:
    properties:
      eventTime:
        type: string
      eventType:
        type: string
      inputs:
        items:
          $ref: '#/definitions/Dataset'
        type: array
      job:
        $ref: '#/definitions/Job'
      outputs:
        items:
          $ref: '#/definitions/Dataset'
        type: array
      producer:
        type: string
      run:
        $ref: '#/definitions/LineageRun'
      schemaURL:
        type: string
    type: object
...
info:
  contact: {}
  description: API for interacting with Marmot
  license:
    name: MIT
    url: https://opensource.org/license/MIT
  title: Marmot API
  version: "0.1"
paths:
  /admin/search/reindex:
    get:
      description: Check whether a search reindex is currently running and whether
        Elasticsearch is configured.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ReindexStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get reindex status
      tags:
      - admin
    post:
      description: Trigger a full reindex from PostgreSQL to Elasticsearch. The reindex
        runs asynchronously in the background. Only one reindex can run at a time.
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/ReindexAcceptedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Start search reindex
      tags:
      - admin
...
  /api/v1/lineage:
    post:
      consumes:
      - application/json
      description: Process OpenLineage run events and update assets/lineage accordingly
      parameters:
      - description: OpenLineage run event
        in: body
        name: event
        required: true
        schema:
          $ref: '#/definitions/RunEvent'
      produces:
      - application/json
      responses:
        "200":
          description: Event processed successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Ingest OpenLineage event
      tags:
      - lineage
...
  /lineage/assets/{id}:
    get:
      consumes:
      - application/json
      description: Get upstream and downstream lineage for a specific asset
      parameters:
      - description: Asset ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Maximum depth of lineage graph
        in: query
        name: limit
        type: integer
      - default: both
        description: Direction of lineage (upstream, downstream, or both)
        enum:
        - upstream
        - downstream
        - both
        in: query
        name: direction
        type: string
      - description: Comma separated edge types to leave out, for example CONTAINS
          to see data flow without structure
        in: query
        name: exclude_types
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LineageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get asset lineage
      tags:
      - lineage
  /lineage/batch:
    post:
      consumes:
      - application/json
      description: Create lineage edges in batch
      parameters:
      - description: Array of lineage edges to create
        in: body
        name: edges
        required: true
        schema:
          items:
            $ref: '#/definitions/LineageEdge'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/BatchLineageResult'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Batch create lineage edges
      tags:
      - lineage
  /lineage/direct:
    post:
      consumes:
      - application/json
      description: Create a direct lineage connection between two assets and returns
        the created edge
      parameters:
      - description: Lineage edge to create
        in: body
        name: edge
        required: true
        schema:
          $ref: '#/definitions/LineageEdge'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LineageEdge'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Create direct lineage
      tags:
      - lineage
  /lineage/direct/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a direct lineage connection by its ID
      parameters:
      - description: Edge ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Delete direct lineage
      tags:
      - lineage
    get:
      consumes:
      - application/json
      description: Get a specific direct lineage connection by its ID
      parameters:
      - description: Edge ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LineageEdge'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      - BearerAuth: []
      summary: Get direct lineage by ID
      tags:
      - lineage
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"